fix(orders): warn before submitting orders on a closed exchange
CreateOrderPage previously had no closed-market awareness — the backend
rejected market orders with "exchange is closed", and the user only saw
that as a generic submit error. Now the form fetches the exchanges list
alongside the security/accounts, looks up the matching exchange by
acronym, and:

- shows a warning banner explaining the exchange is closed (or open via
  override, weekend, outside hours) and that Limit/Stop orders are still
  fine but Market orders will be rejected (TestoviCelina3 §S45/§S46);
- pre-flight blocks market orders with an inline error so the user gets
  a fast, clear message instead of a backend round-trip.

The open/closed computation duplicated between BerzaPage and the new
order-form check moves to src/utils/exchangeHours.js so both stay in
sync. BerzaPage now imports computeExchangeStatus from there;
CreateOrderPage imports findExchangeByAcronym + computeExchangeStatus.

For testing outside trading hours: flip the per-exchange "Test mod"
toggle on /berze (open_override). The shared helper honors override the
same way the backend does. After-hours simulation is intentionally not
covered by this change — see e2e.txt:19 + hours.go:134.
